# Replace Struct with internal repeated KeyValue