--- a/input/elasticapm/internal/modeldecoder/modeldecoderutil/keyvaluepb.go
+++ b/input/elasticapm/internal/modeldecoder/modeldecoderutil/keyvaluepb.go
@@ -15,38 +15,23 @@
 // specific language governing permissions and limitations
 // under the License.
 
-syntax = "proto3";
+package modeldecoderutil
 
-package elastic.apm.v1;
+import "github.com/elastic/apm-data/model/modelpb"
 
-import "google/protobuf/struct.proto";
-import "headers.proto";
-import "keyvalue.proto";
+func ToKv(m map[string]any) []*modelpb.KeyValue {
+	nm := normalizeMap(m)
+	if len(nm) == 0 {
+		return nil
+	}
 
-option go_package = "github.com/elastic/apm-data/model/modelpb";
+	kv := []*modelpb.KeyValue{}
+	for k, v := range m {
+		kv = append(kv, &modelpb.KeyValue{
+			Key:   k,
+			Value: v.(string),
+		})
+	}
 
-message HTTP {
-  HTTPRequest request = 1;
-  HTTPResponse response = 2;
-  string version = 3;
-}
-
-message HTTPRequest {
-  google.protobuf.Value body = 1;
-  repeated HTTPHeader headers = 2;
-  repeated KeyValue env = 3;
-  repeated KeyValue cookies = 4;
-  string id = 5;
-  string method = 6;
-  string referrer = 7;
-}
-
-message HTTPResponse {
-  repeated HTTPHeader headers = 1;
-  optional bool finished = 2;
-  optional bool headers_sent = 3;
-  optional int64 transfer_size = 4;
-  optional int64 encoded_body_size = 5;
-  optional int64 decoded_body_size = 6;
-  int32 status_code = 7;
+	return kv
 }