--- a/input/elasticapm/internal/modeldecoder/rumv3/decoder.go
+++ b/input/elasticapm/internal/modeldecoder/rumv3/decoder.go
@@ -209,7 +209,7 @@ func mapToErrorModel(from *errorEvent, event *modelpb.APMEvent) {
 			}
 		}
 		if len(from.Context.Custom) > 0 {
-			out.Custom = modeldecoderutil.ToStruct(from.Context.Custom)
+			out.Custom = modeldecoderutil.ToKv(from.Context.Custom)
 		}
 	}
 	if from.Culprit.IsSet() {
@@ -275,7 +275,7 @@ func mapToErrorModel(from *errorEvent, event *modelpb.APMEvent) {
 
 func mapToExceptionModel(from errorException, out *modelpb.Exception) {
 	if len(from.Attributes) > 0 {
-		out.Attributes = modeldecoderutil.ToStruct(from.Attributes)
+		out.Attributes = modeldecoderutil.ToKv(from.Attributes)
 	}
 	if from.Code.IsSet() {
 		out.Code = modeldecoderutil.ExceptionCodeString(from.Code.Val)
@@ -447,7 +447,7 @@ func mapToRequestModel(from contextRequest, out *modelpb.HTTPRequest) {
 		out.Method = from.Method.Val
 	}
 	if len(from.Env) > 0 {
-		out.Env = modeldecoderutil.ToStruct(from.Env)
+		out.Env = modeldecoderutil.ToKv(from.Env)
 	}
 	if from.Headers.IsSet() {
 		out.Headers = modeldecoderutil.HTTPHeadersToModelpb(from.Headers.Val)
@@ -701,7 +701,7 @@ func mapToTransactionModel(from *transaction, event *modelpb.APMEvent) {
 	// map transaction specific data
 	if from.Context.IsSet() {
 		if len(from.Context.Custom) > 0 {
-			out.Custom = modeldecoderutil.ToStruct(from.Context.Custom)
+			out.Custom = modeldecoderutil.ToKv(from.Context.Custom)
 		}
 		if len(from.Context.Tags) > 0 {
 			modeldecoderutil.MergeLabels(from.Context.Tags, event)