--- a/input/elasticapm/internal/modeldecoder/v2/decoder.go
+++ b/input/elasticapm/internal/modeldecoder/v2/decoder.go
@@ -434,7 +434,7 @@ func mapToErrorModel(from *errorEvent, event *modelpb.APMEvent) {
 			}
 		}
 		if len(from.Context.Custom) > 0 {
-			out.Custom = modeldecoderutil.ToStruct(from.Context.Custom)
+			out.Custom = modeldecoderutil.ToKv(from.Context.Custom)
 		}
 	}
 	if from.Culprit.IsSet() {
@@ -499,7 +499,7 @@ func mapToErrorModel(from *errorEvent, event *modelpb.APMEvent) {
 
 func mapToExceptionModel(from errorException, out *modelpb.Exception) {
 	if len(from.Attributes) > 0 {
-		out.Attributes = modeldecoderutil.ToStruct(from.Attributes)
+		out.Attributes = modeldecoderutil.ToKv(from.Attributes)
 	}
 	if from.Code.IsSet() {
 		out.Code = modeldecoderutil.ExceptionCodeString(from.Code.Val)
@@ -813,13 +813,13 @@ func mapToRequestModel(from contextRequest, out *modelpb.HTTPRequest) {
 		out.Method = from.Method.Val
 	}
 	if len(from.Env) > 0 {
-		out.Env = modeldecoderutil.ToStruct(from.Env)
+		out.Env = modeldecoderutil.ToKv(from.Env)
 	}
 	if from.Body.IsSet() {
 		out.Body = modeldecoderutil.ToValue(modeldecoderutil.NormalizeHTTPRequestBody(from.Body.Val))
 	}
 	if len(from.Cookies) > 0 {
-		out.Cookies = modeldecoderutil.ToStruct(from.Cookies)
+		out.Cookies = modeldecoderutil.ToKv(from.Cookies)
 	}
 	if from.Headers.IsSet() {
 		out.Headers = modeldecoderutil.HTTPHeadersToModelpb(from.Headers.Val)
@@ -1253,7 +1253,7 @@ func mapToStracktraceModel(from []stacktraceFrame, out []*modelpb.StacktraceFram
 			copy(fr.PreContext, eventFrame.PreContext)
 		}
 		if len(eventFrame.Vars) > 0 {
-			fr.Vars = modeldecoderutil.ToStruct(eventFrame.Vars)
+			fr.Vars = modeldecoderutil.ToKv(eventFrame.Vars)
 		}
 		out[idx] = &fr
 	}
@@ -1284,7 +1284,7 @@ func mapToTransactionModel(from *transaction, event *modelpb.APMEvent) {
 
 	if from.Context.IsSet() {
 		if len(from.Context.Custom) > 0 {
-			out.Custom = modeldecoderutil.ToStruct(from.Context.Custom)
+			out.Custom = modeldecoderutil.ToKv(from.Context.Custom)
 		}
 		if len(from.Context.Tags) > 0 {
 			modeldecoderutil.MergeLabels(from.Context.Tags, event)