--- a/model/modelpb/error.pb.json.go
+++ b/model/modelpb/error.pb.json.go
@@ -29,7 +29,7 @@ func (e *Error) toModelJSON(out *modeljson.Error) {
 		StackTrace:  e.StackTrace,
 	}
 	if e.Custom != nil {
-		m := e.Custom.AsMap()
+		m := kvToMap(e.Custom)
 		updateFields(m)
 		out.Custom = m
 	}
@@ -64,7 +64,7 @@ func (e *Exception) toModelJSON(out *modeljson.Exception) {
 		Handled: e.Handled,
 	}
 	if e.Attributes != nil {
-		out.Attributes = e.Attributes.AsMap()
+		out.Attributes = kvToMap(e.Attributes)
 	}
 	if n := len(e.Cause); n > 0 {
 		out.Cause = make([]modeljson.Exception, n)