--- a/model/modelpb/error.pb.json_test.go
+++ b/model/modelpb/error.pb.json_test.go
@@ -28,7 +28,7 @@ import (
 func TestErrorToModelJSON(t *testing.T) {
 	handled := true
 
-	attrs, attrsMap := randomStruct(t)
+	attrs, attrsMap := randomKv(t)
 
 	testCases := map[string]struct {
 		proto    *Error