… codec/fullevent_test.go
… codec/utils_test.go
… model/modelpb/apmevent.pb.json_test.go
… model/modelpb/error.pb.go
… model/modelpb/http.pb.go
… model/modelpb/http.pb.json_test.go
… model/modelpb/keyvalue.pb.go
… model/modelpb/stacktrace.pb.go
… model/modelpb/stacktrace.pb.json_test.go
… model/modelpb/transaction.pb.go
… model/modelpb/transaction_vtproto.pb.go
… model/modelpb/util_test.go
… model/proto/stacktrace.proto
… model/proto/transaction.proto