As described in https://github.com/elastic/apm-data/issues/47#issuecomment-1595599688