@dmathieu 2023-07-17
I must admit that I'm a bit surprised to see such a difference in the benchmarks. It's a good one though!

```
Running benchstat
goos: darwin
goarch: arm64
pkg: github.com/elastic/apm-data/codec
                         │ bench-main  │             bench-branch             │
                         │   sec/op    │    sec/op     vs base                │
Encode/format=json-10      21.77µ ± 5%   21.93µ ±  2%        ~ (p=1.000 n=10)
Encode/format=vtproto-10   2.475µ ± 2%   1.071µ ±  2%  -56.73% (p=0.000 n=10)
Decode/format=json-10      17.87µ ± 4%   16.44µ ± 12%        ~ (p=0.063 n=10)
Decode/format=vtproto-10   6.908µ ± 2%   6.200µ ±  5%  -10.26% (p=0.000 n=10)
geomean                    9.031µ        6.994µ        -22.55%

                         │  bench-main  │             bench-branch              │
                         │   bytes/op   │   bytes/op    vs base                 │
Encode/format=json-10      5.988Ki ± 0%   5.988Ki ± 0%       ~ (p=1.000 n=10) ¹
Encode/format=vtproto-10   2.726Ki ± 0%   2.710Ki ± 0%  -0.57% (p=0.000 n=10)
geomean                    4.040Ki        4.028Ki       -0.29%
¹ all samples are equal

pkg: github.com/elastic/apm-data/input/elasticapm/internal/decoder
                   │  bench-main  │          bench-branch          │
                   │    sec/op    │    sec/op     vs base          │
NDStreamDecoder-10   42.26n ± 10%   45.00n ± 25%  ~ (p=0.190 n=10)

pkg: github.com/elastic/apm-data/model/modelpb
                          │ bench-main  │         bench-branch          │
                          │   sec/op    │   sec/op     vs base          │
APMEventToJSON/to-json-10   11.55µ ± 3%   11.40µ ± 7%  ~ (p=0.579 n=10)
```